% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_databox}
\alias{run_databox}
\title{Run the databoxR Shiny Application}
\usage{
run_databox(demo_data = FALSE, ...)
}
\arguments{
\item{demo_data}{Logical, if TRUE loads sample datasets automatically. Default is FALSE.}

\item{...}{arguments to pass to shinyApp}
}
\value{
A shiny application object
}
\description{
This function launches the main databoxR application for exploring SDTM/ADaM datasets
}
\examples{
\dontrun{
  run_databox()
  run_databox(demo_data = TRUE)  # Load with sample data
}
}